backend/requirements.txt
fastapi
uvicorn
google-genai>=1.1.0
